testing/orchestrator/requirements.txt
# ====================================================================
# AI-Code 测试编排器依赖包
# 兼容 Python 3.11+ (包括 Python 3.13)
# ====================================================================

# 核心必需依赖
typer[all]==0.12.5              # CLI 框架（包含 rich 支持）
rich==13.7.1                   # 丰富的控制台输出和表格
pyyaml==6.0.1                  # YAML 配置文件支持
psutil==5.9.8                  # 系统和进程监控
aiohttp==3.9.1                 # 异步 HTTP 客户端
jinja2==3.1.3                  # 模板引擎（报告生成）

# 文件和路径处理
watchdog==4.0.0               # 文件系统监控（watch 模式）

# 异步和并发
aiofiles==23.2.1              # 异步文件操作

# Git 集成
GitPython==3.1.41             # Git 操作和差异检测

# HTTP 客户端（健康检查）
httpx==0.26.0                 # 现代异步 HTTP 客户端
requests==2.31.0              # 传统 HTTP 客户端（兼容性）

# 数据处理和序列化
orjson==3.9.15                # 快速 JSON 处理

# 日志系统
loguru==0.7.2                 # 高级日志库
colorlog==6.8.2               # 彩色日志输出

# 环境变量和配置
python-dotenv==1.0.0          # 环境变量管理
python-decouple==3.8          # 配置管理

# 时间处理
python-dateutil==2.8.2        # 日期时间处理
schedule==1.2.0               # 任务调度

# 实用工具
tqdm==4.66.2                  # 进度条
tabulate==0.9.0               # 表格输出
humanize==4.9.0               # 人性化显示

# 测试和质量保证
pytest==8.0.0                 # Python 测试框架
coverage==7.4.1               # 覆盖率工具
pytest-html==4.1.1            # HTML 测试报告
pytest-asyncio==0.23.5        # 异步测试支持
pytest-mock==3.12.0           # 模拟对象

# 代码质量工具
black==24.1.1                 # 代码格式化
isort==5.13.2                 # import 排序
mypy==1.8.0                   # 类型检查
bandit==1.7.6                 # 安全扫描

# 调试和开发工具
icecream==2.1.3               # 调试工具

# 报告生成和可视化
plotly==5.18.0                # 图表生成
pandas==2.2.0                 # 数据处理和分析
markdown==3.5.2               # Markdown 支持

# ====================================================================
# 可选依赖 (按需安装)
# 注意：以下依赖可能在 Python 3.13 上有兼容性问题
# ====================================================================

# 数据验证（可选 - 目前与 Python 3.13 有兼容性问题）
# pydantic==2.6.1               # 数据验证和序列化 - 暂时跳过

# 通知系统（可选）
# slack-sdk==3.26.2             # Slack 通知
# python-telegram-bot==20.8     # Telegram 通知

# 数据库支持（可选）
# sqlalchemy==2.0.25            # ORM
# redis==5.0.1                  # Redis 缓存
# psycopg2-binary==2.9.9        # PostgreSQL 驱动

# 系统集成（可选）
# docker==7.0.0                 # Docker API 客户端

# CI/CD 集成（可选）
# github3.py==4.0.1             # GitHub API 客户端
# python-gitlab==4.4.0          # GitLab API 客户端

# 质量门禁（可选）
# sonarqube-api==1.3.4          # SonarQube 集成
# codecov==2.1.13               # Codecov 集成

# Web 界面（可选）
# fastapi==0.109.0              # Web API 框架 - 依赖 pydantic
# uvicorn[standard]==0.27.0     # ASGI 服务器
# websockets==12.0              # WebSocket 支持

# 性能监控（可选）
# memory-profiler==0.61.0       # 内存使用分析

# 安全和加密（可选）
# cryptography==42.0.2          # 加密库
# keyring==24.3.0               # 凭证存储
# safety==3.0.1                 # 依赖安全检查

# 缓存（可选）
# diskcache==5.6.3              # 磁盘缓存
# joblib==1.3.2                 # 内存缓存和并行

# 文件处理（可选）
# openpyxl==3.1.2               # Excel 文件处理
# lxml==5.1.0                   # XML/HTML 解析
# beautifulsoup4==4.12.3        # HTML 解析

# 测试数据生成（可选）
# faker==22.7.0                 # 测试数据生成
# factory-boy==3.3.0            # 对象工厂
# responses==0.25.0             # HTTP 响应模拟

# ====================================================================
# Python 3.13 兼容性说明
# ====================================================================
# 
# 目前已知与 Python 3.13 不兼容的包：
# - pydantic (依赖 pydantic-core，Rust 编译问题)
# - fastapi (依赖 pydantic)
# 
# 这些包将在未来版本中支持 Python 3.13
# 当前版本的测试编排器不依赖这些包的核心功能
# 
# ====================================================================

# ====================================================================
# 安装说明
# ====================================================================
# 
# 基础安装（推荐）:
# pip install -r requirements.txt
#
# 如需 pydantic 功能，请使用 Python 3.12:
# pyenv install 3.12.7
# pyenv local 3.12.7
# pip install pydantic==2.6.1
#
# ====================================================================